Fix invalid cert and access token object
Added an ‘InvalidCert” option to let user decide what to do when the
certificate received from server is a self signed certificate, when set
to true, the certificate error is ignored  and response is processed;
when set to false (default), the code looks at certificates on the
user computer to see if the certificate is stored locally, if it is,
the response is processed, if not, an exception will be thrown.

Changed login URL and the time interval for the access token refresh
rate. Changed the AccessToken object that was required to convert
the login credentials into the object.